Added Get VaultKeeps by User

+++ Repositories/VaultKeepsRepository.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Data;
 using Dapper;
 using keepr.Models;
+using Keepr.Models;
 
 namespace keepr.Repositories
 {
@@ -13,24 +14,36 @@ namespace keepr.Repositories
     {
       _db = db;
     }
+    internal IEnumerable<DTOVaultKeep> GetVaultKeepsByUser(string userId)
+    {
+      string sql = "SELECT * FROM vaultkeeps WHERE userId = @userId";
+      return _db.Query<DTOVaultKeep>(sql, new{ userId });
+    }
     internal DTOVaultKeep GetById(int Id)
     {
       string sql = "SELECT * FROM vaultkeeps WHERE id = @Id";
       return _db.QueryFirstOrDefault<DTOVaultKeep>(sql, new { Id });
     }
 
-    internal IEnumerable<VaultKeepViewModel> GetByUser(string userId)
-    {
-      string sql = @"
-            SELECT
-            k.*,
-            vk.id as vaultKeepId
-            FROM vaultkeeps vk
-            INNER JOIN vaults v ON v.id = vk.vaultId
-            WHERE userId = @userId;";
+       internal IEnumerable<VaultKeepViewModel> GetAll()
+       {
+         string sql = "SELECT * FROM vaultkeeps";
+         return _db.Query<VaultKeepViewModel>(sql);
+       }
 
-            return _db.Query<VaultKeepViewModel>(sql, new { userId });
-    }
+
+    // internal IEnumerable<VaultKeepViewModel> GetByUser()
+    // {
+    //   string sql = @"
+    //         SELECT
+    //         k.*,
+    //         vk.id as vaultKeepId
+    //         FROM vaultkeeps vk
+    //         INNER JOIN vaults v ON v.id = vk.vaultId
+    //         WHERE userId = @UserId;";
+
+    //         return _db.Query<VaultKeepViewModel>(sql);
+    // }
 
     internal int Create(DTOVaultKeep newDTOVaultKeep)
     {
@@ -43,13 +56,14 @@ namespace keepr.Repositories
       return _db.ExecuteScalar<int>(sql, newDTOVaultKeep);
     }
 
-    internal void Delete(int Id)
+    internal bool Delete(int Id)
     {
-      string sql = "DELETE FROM vaultkeeps WHERE id = @Id";
-      _db.Execute(sql, new { Id });
+      string sql = "DELETE FROM vaultkeeps WHERE id = @Id AND userId = @UserId LIMIT 1";
+      int affectedRows = _db.Execute(sql, new {Id});
+      return affectedRows == 1;
     }
 
-    internal IEnumerable<VaultKeepViewModel> GetKeepsByVaultId(int id)
+    internal IEnumerable<VaultKeepViewModel> GetKeepsByVaultId(int id, string userId)
     {
       string sql = @"
            SELECT 
@@ -58,7 +72,7 @@ namespace keepr.Repositories
             FROM vaultkeeps vk
             INNER JOIN keeps k ON k.id = vk.keepId 
             WHERE (vaultId = @vaultId AND vk.userId = @userId)";
-          return _db.Query<VaultKeepViewModel>(sql, new { id });
+          return _db.Query<VaultKeepViewModel>(sql, new { id, userId });
     }
   }
 }
